requirements.txt
requests
Click
python-dotenv